[ fix ] Everyone should be able to build the pages
Everyone should be able to build the pages locally without the source
markdown files being affected. In the previous configuration, links
to the index were added to each one of the .md files thus polluting
the whole thing. We now only ever modify copies of the versioned files
thus keeping the repo in a sane state.